feat: Separathe warehouse locator.
