+++ config/default.json
@@ -279,6 +279,7 @@
         "adempiere/general-ledger",
         "adempiere/general-ledger/accouting-combination",
         "adempiere/material-management",
+        "adempiere/material-management/warehouse-locator",
         "adempiere/user",
         "adempiere/user/enrollment",
         "adempiere/user/log",
