+++ src/modules/adempiere-api/api/extensions/adempiere/material-management/index.js
@@ -14,13 +14,12 @@
  ************************************************************************************/
 
 import { Router } from 'express';
+
 import { convertEntitiesListFromGRPC } from '../util/convertData';
 import {
   convertProductAttributeSet,
   convertProductAttributeSetInstance,
-  convertListProductAttributeSetInstances,
-  convertListAvailableWarehouses,
-  convertListLocators
+  convertListProductAttributeSetInstances
 } from '../util/convertMaterialManagement';
 
 module.exports = ({ config }) => {
@@ -208,89 +207,6 @@ module.exports = ({ config }) => {
           res.json({
             code: 200,
             result: convertProductAttributeSetInstance(response)
-          })
-        } else if (err) {
-          res.json({
-            code: 500,
-            result: err.details
-          })
-        }
-      })
-    }
-  });
-
-  /**
-   * POST List Available Warehouses
-   *
-   * req.query.token - user token
-   * req.query.language - login language
-   * req.query.search_value - search value optional
-   * req.body.warehouse_id
-   * req.body.warehouse_uuid
-   */
-  api.post('/list-available-warehouses', (req, res) => {
-    if (req.body) {
-      service.listAvailableWarehouses({
-        token: req.query.token,
-        language: req.query.language,
-        // DSL Query
-        searchValue: req.query.search_value,
-        warehouseId: req.body.warehouse_id,
-        warehouseUuid: req.body.warehouse_uuid,
-        // Page Data
-        pageSize: req.query.page_size,
-        pageToken: req.query.page_token
-      }, (err, response) => {
-        if (response) {
-          res.json({
-            code: 200,
-            result: convertListAvailableWarehouses(response)
-          })
-        } else if (err) {
-          res.json({
-            code: 500,
-            result: err.details
-          })
-        }
-      })
-    }
-  });
-
-  /**
-   * POST List Warehouses Locators
-   *
-   * req.query.token - user token
-   * req.query.language - login language
-   * req.query.search_value - search value optional
-   * req.body.warehouse_id
-   * req.body.warehouse_uuid
-   */
-  api.post('/list-warehouse-locators', (req, res) => {
-    if (req.body) {
-      service.listLocators({
-        token: req.query.token,
-        language: req.query.language,
-        // DSL Query
-        searchValue: req.query.search_value,
-        warehouseId: req.body.warehouse_id,
-        warehouseUuid: req.body.warehouse_uuid,
-        contextAttributes: req.body.context_attributes,
-        // Dynamic Validation Query
-        fieldUuid: req.body.field_uuid,
-        processParameterUuid: req.body.process_parameter_uuid,
-        browseFieldUuid: req.body.browse_field_uuid,
-        referenceUuid: req.body.reference_uuid,
-        columnUuid: req.body.column_uuid,
-        tableName: req.body.table_name,
-        columnName: req.body.column_name,
-        // Page Data
-        pageSize: req.query.page_size,
-        pageToken: req.query.page_token
-      }, (err, response) => {
-        if (response) {
-          res.json({
-            code: 200,
-            result: convertListLocators(response)
           })
         } else if (err) {
           res.json({
